flowedit: Extract canvas message handling from main.rs (#2593)
Move all CanvasMessage handlers plus ZoomIn/ZoomOut/ToggleAutoFit from
FlowEdit::update() to canvas_view.rs. Adds CanvasAction enum for
cross-module effects (OpenNode). main.rs reduced by ~240 lines.
